+++ tests/SkipCheckerTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace MyTester;
 
+use MyTester\Attributes\IgnoreDeprecations;
 use MyTester\Attributes\Skip;
 use MyTester\Attributes\TestSuite;
 
@@ -19,37 +20,49 @@ final class SkipCheckerTest extends TestCase
         return (new SkipChecker($this->annotationsReader));
     }
 
+    #[IgnoreDeprecations]
     public function testCheckPhpVersion(): void
     {
-        $this->assertNull($this->getSkipChecker()->checkPhpVersion(1));
+        $this->assertNull($this->getSkipChecker()->checkPhpVersion(1)); // @phpstan-ignore method.deprecated
+        // @phpstan-ignore method.deprecated
         $this->assertType("string", $this->getSkipChecker()->checkPhpVersion(PHP_INT_MAX));
     }
 
+    #[IgnoreDeprecations]
     public function testCheckLoadedExtension(): void
     {
-        $this->assertNull($this->getSkipChecker()->checkLoadedExtension("ctype"));
+        $this->assertNull($this->getSkipChecker()->checkLoadedExtension("ctype")); // @phpstan-ignore method.deprecated
+        // @phpstan-ignore method.deprecated
         $this->assertType("string", $this->getSkipChecker()->checkLoadedExtension("abc"));
     }
 
+    #[IgnoreDeprecations]
     public function testCheckPhpSapi(): void
     {
-        $this->assertNull($this->getSkipChecker()->checkPhpSapi(PHP_SAPI));
-        $this->assertType("string", $this->getSkipChecker()->checkPhpSapi("abc"));
+        $this->assertNull($this->getSkipChecker()->checkPhpSapi(PHP_SAPI)); // @phpstan-ignore method.deprecated
+        $this->assertType("string", $this->getSkipChecker()->checkPhpSapi("abc")); // @phpstan-ignore method.deprecated
     }
 
+    #[IgnoreDeprecations]
     public function testCheckOsFamily(): void
     {
-        $this->assertNull($this->getSkipChecker()->checkOsFamily(PHP_OS_FAMILY));
+        $this->assertNull($this->getSkipChecker()->checkOsFamily(PHP_OS_FAMILY)); // @phpstan-ignore method.deprecated
+        // @phpstan-ignore method.deprecated
         $this->assertType("string", $this->getSkipChecker()->checkOsFamily("Solaris"));
     }
 
+    #[IgnoreDeprecations]
     public function testGetSkipValue(): void
     {
+        // @phpstan-ignore method.deprecated
         $this->assertNull($this->getSkipChecker()->getSkipValue(static::class, "skipNull"));
+        // @phpstan-ignore method.deprecated
         $this->assertSame([], $this->getSkipChecker()->getSkipValue(static::class, "skip"));
+        // @phpstan-ignore method.deprecated
         $this->assertSame(["php" => 666, ], $this->getSkipChecker()->getSkipValue(static::class, "skipArray"));
     }
 
+    #[IgnoreDeprecations]
     public function testShouldSkip(): void
     {
         $this->assertSame(false, $this->getSkipChecker()->shouldSkip(static::class, "skipNull"));